requirements.txt
torch
torch-geometric
pytorch-lightning
hydra-core
omegaconf
numpy
pandas
matplotlib
tensorboard
